ForceAtlas2: continue #5213 for `prevent_overlapping` and `mobility`
This PR is a continuation of #5213 by @MathisHammel and aims to address issues
that were identified in that PR and pay off some technical debt.

**This PR is a work in progress**

Some changes:
- Add support for passing vertex ids for initial positions, vertex_radius, and mobility.
- Add support for passing random_state.
- Add support int64_t vertex/edge types.
- Remove support for `callback=` argument in `cugraph.force_atlas2` (it now gives informative error).
- Remove `cugraph/layout/force_atlas2_wrapper.pyx` (use PLC instead).
- Fix mutating input arrays (we now copy and use `const` in C++).
- Fix inputs and result when integral vertex ids weren't contiguous.
  - For example, if vertices are `[0, 2]`, inputs and result would contain values for `1`.
- Fixing renumbering of vertices.

Many things are still experimental and broken, but I am continuing to experiment and fix things :)

+++ cpp/src/layout/legacy/exact_repulsion.cuh
@@ -79,8 +79,8 @@ void apply_repulsion(const float* restrict x_pos,
                      cudaStream_t stream)
 {
   dim3 nthreads(TPB_X, TPB_Y);
-  dim3 nblocks(min((n + nthreads.x - 1) / nthreads.x, CUDA_MAX_BLOCKS_2D),
-               min((n + nthreads.y - 1) / nthreads.y, CUDA_MAX_BLOCKS_2D));
+  dim3 nblocks(min((static_cast<int>(n) + nthreads.x - 1) / nthreads.x, CUDA_MAX_BLOCKS_2D),
+               min((static_cast<int>(n) + nthreads.y - 1) / nthreads.y, CUDA_MAX_BLOCKS_2D));
 
   repulsion_kernel<vertex_t><<<nblocks, nthreads, 0, stream>>>(x_pos,
                                                                y_pos,
